+++ packages/effects-core/src/math/translate.ts
@@ -1,7 +1,6 @@
 import { Euler } from '@galacean/effects-math/es/core/euler';
 import { Matrix4 } from '@galacean/effects-math/es/core/matrix4';
 import { Vector3 } from '@galacean/effects-math/es/core/vector3';
-import type { ItemLinearVelOverLifetime } from '../plugins';
 import type { ValueGetter } from './value-getters';
 
 export function translatePoint (x: number, y: number): number[] {
@@ -17,6 +16,14 @@ export function translatePoint (x: number, y: number): number[] {
 
 const tempEuler = new Euler();
 const tempMat4 = new Matrix4();
+
+export type ItemLinearVelOverLifetime = {
+  asMovement?: boolean,
+  x?: ValueGetter<number>,
+  y?: ValueGetter<number>,
+  z?: ValueGetter<number>,
+  enabled?: boolean,
+};
 
 export interface TranslateTarget {
   speedOverLifetime?: ValueGetter<number>,
