--- packages/effects-core/src/plugins/cal/animation-playable.ts
@@ -1,11 +1,0 @@
-import { AnimationStream } from './animation-stream';
-import { Playable } from './playable-graph';
-
-export class AnimationPlayable extends Playable {
-  animationStream: AnimationStream;
-
-  constructor () {
-    super();
-    this.animationStream = new AnimationStream(this);
-  }
-}
