+++ packages/effects-core/src/plugins/particle/particle-vfx-item.ts
@@ -1,6 +1,6 @@
 import { VFXItem } from '../../vfx-item';
-import type { FrameContext } from '../cal/playable-graph';
-import { Playable, PlayableAsset } from '../cal/playable-graph';
+import type { FrameContext } from '../timeline/playable';
+import { Playable, PlayableAsset } from '../timeline/playable';
 import { ParticleSystem } from './particle-system';
 
 /**
