+++ packages/effects-core/src/plugins/sprite/sprite-item.ts
@@ -6,8 +6,8 @@ import { BaseRenderComponent, EffectComponent } from '../../components';
 import { effectsClass } from '../../decorators';
 import type { Engine } from '../../engine';
 import { TextureSourceType, type Texture2DSourceOptionsVideo } from '../../texture';
-import type { FrameContext } from '../cal/playable-graph';
-import { Playable, PlayableAsset } from '../cal/playable-graph';
+import type { FrameContext } from '../timeline/playable';
+import { Playable, PlayableAsset } from '../timeline/playable';
 import { TrackAsset } from '../timeline/track';
 import { TrackMixerPlayable } from '../timeline/playables/track-mixer-playable';
 import type { VFXItem } from '../../vfx-item';
