+++ packages/effects-core/src/plugins/timeline/index.ts
@@ -2,3 +2,4 @@ export * from './track';
 export * from './tracks';
 export * from './playable-assets';
 export * from './timeline-asset';
+export * from './playable';
