+++ packages/effects-core/src/plugins/timeline/playable-assets/sub-composition-playable-asset.ts
@@ -1,7 +1,7 @@
 import * as spec from '@galacean/effects-specification';
 import { effectsClass } from '../../../decorators';
-import type { Playable } from '../../cal/playable-graph';
-import { PlayableAsset } from '../../cal/playable-graph';
+import type { Playable } from '../playable';
+import { PlayableAsset } from '../playable';
 import { SubCompositionClipPlayable } from '../playables';
 
 @effectsClass(spec.DataType.SubCompositionPlayableAsset)
