+++ packages/effects-core/src/plugins/timeline/playables/activation-mixer-playable.ts
@@ -1,5 +1,5 @@
 import { VFXItem } from '../../../vfx-item';
-import type { FrameContext } from '../../cal/playable-graph';
+import type { FrameContext } from '../playable';
 import { TrackMixerPlayable } from './track-mixer-playable';
 
 export class ActivationMixerPlayable extends TrackMixerPlayable {
