+++ packages/effects-core/src/plugins/timeline/timeline-asset.ts
@@ -2,8 +2,8 @@ import * as spec from '@galacean/effects-specification';
 import { effectsClass, serialize } from '../../decorators';
 import type { VFXItem } from '../../vfx-item';
 import type { RuntimeClip, TrackAsset } from './track';
-import { ObjectBindingTrack } from '../cal/calculate-item';
-import { PlayState } from '../cal/playable-graph';
+import { ObjectBindingTrack } from './tracks/object-binding-track';
+import { PlayState } from './playable';
 import type { Constructor } from '../../utils';
 import { TrackInstance } from './track-instance';
 import type { SceneBinding } from 'packages/effects-core/src/comp-vfx-item';
