+++ packages/effects-core/src/plugins/timeline/track-instance.ts
@@ -1,4 +1,4 @@
-import type { PlayableOutput } from '../cal/playable-graph';
+import type { PlayableOutput } from './playable';
 import type { TrackMixerPlayable } from './playables';
 import type { TrackAsset } from './track';
