+++ packages/effects-core/src/plugins/timeline/tracks/index.ts
@@ -8,3 +8,4 @@ export * from './particle-track';
 export * from './property-track';
 export * from './vector-property-track';
 export * from './color-property-track';
+export * from './object-binding-track';
